# demucs_service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Demucs Service API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://demucs.stanley.arpa"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/status": {
      "get": {
        "summary": "Get service status",
        "responses": {
          "200": {
            "description": "Service status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/clear-caches": {
      "post": {
        "summary": "Clear all cached artifacts and all jobs",
        "responses": {
          "200": {
            "description": "Caches cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearCachesResponse"
                }
              }
            }
          },
          "409": {
            "description": "Jobs are still running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/models": {
      "get": {
        "summary": "List available models",
        "responses": {
          "200": {
            "description": "Models list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "List jobs",
        "description": "Returns jobs ordered by creation time descending. Queued jobs include `queue_position` computed in FIFO order (1 = next to run).",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "running",
                "succeeded",
                "failed"
              ]
            },
            "description": "Filter by job status"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Max jobs to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a separation job (exactly one file)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "mode",
                  "model",
                  "file"
                ],
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": [
                      "4",
                      "2",
                      "both"
                    ]
                  },
                  "model": {
                    "type": "string",
                    "description": "Demucs model name"
                  },
                  "job_label": {
                    "type": "string",
                    "description": "Optional display/log label only; never used for identity, dedupe, cache keys, or output paths."
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Single MP3 file. Exactly one file is supported per job."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "multi_file_not_supported": {
                    "summary": "More than one file uploaded",
                    "value": {
                      "error": "invalid_request",
                      "message": "Exactly one file is supported per job; multi-file uploads are not supported."
                    }
                  },
                  "invalid_mode": {
                    "summary": "Unsupported mode value",
                    "value": {
                      "error": "invalid_mode",
                      "message": "Invalid mode."
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unsupported_media_type": {
                    "summary": "Wrong file type",
                    "value": {
                      "error": "unsupported_media_type",
                      "message": "Only mp3 files supported: track.wav"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Creates exactly one asynchronous job from exactly one uploaded MP3 file. If zero or multiple files are provided in the multipart payload, the request is rejected and no partial work is started."
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "Get job details",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/result": {
      "get": {
        "summary": "Download job result zip",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output zip",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "description": "Attachment filename for the ZIP artifact.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Job has not succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "job_not_succeeded": {
                    "summary": "Job not finished yet",
                    "value": {
                      "error": "job_not_succeeded",
                      "message": "Output not ready"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Job or output missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Downloads the output ZIP for a succeeded job. ZIP contains deterministic top-level directories only: `all/` (mode 4 or both) and/or `vocals/` (mode 2 or both)."
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "OpenAPI JSON endpoint",
        "responses": {
          "200": {
            "description": "OpenAPI specification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/docs": {
      "get": {
        "summary": "Interactive API docs",
        "responses": {
          "200": {
            "description": "Swagger UI",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "service": {
            "type": "string",
            "example": "demucs"
          },
          "running_jobs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "max_concurrent_jobs": {
            "type": "integer"
          },
          "cuda": {
            "$ref": "#/components/schemas/CudaStatus"
          },
          "storage_volume": {
            "$ref": "#/components/schemas/StorageVolume"
          }
        },
        "required": [
          "service",
          "running_jobs",
          "max_concurrent_jobs",
          "cuda",
          "storage_volume"
        ]
      },
      "CudaStatus": {
        "type": "object",
        "properties": {
          "cuda_available": {
            "type": "boolean",
            "example": true
          },
          "cuda_device_count": {
            "type": "integer",
            "example": 1
          },
          "cuda_device_name": {
            "type": "string",
            "example": "NVIDIA GeForce RTX 4090"
          },
          "torch_cuda_version": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "cuda_available",
          "cuda_device_count",
          "cuda_device_name"
        ]
      },
      "StorageVolume": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "total_bytes": {
            "type": "integer"
          },
          "used_bytes": {
            "type": "integer"
          },
          "free_bytes": {
            "type": "integer"
          }
        },
        "required": [
          "path",
          "total_bytes",
          "used_bytes",
          "free_bytes"
        ]
      },
      "ClearCachesResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "cleared": {
            "type": "object",
            "properties": {
              "artifacts_root": {
                "type": "string"
              },
              "jobs_root": {
                "type": "string"
              }
            },
            "required": [
              "artifacts_root",
              "jobs_root"
            ]
          }
        },
        "required": [
          "ok",
          "cleared"
        ]
      },
      "ModelsResponse": {
        "type": "object",
        "properties": {
          "default": {
            "type": "string"
          },
          "models": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "default",
          "models"
        ]
      },
      "JobsResponse": {
        "type": "object",
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            }
          }
        },
        "required": [
          "jobs"
        ]
      },
      "CreateJobResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "example": {
          "id": "5e3db5f098324f8f85cd2cb7486b0d59"
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code in snake_case"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "example": {
          "error": "invalid_request",
          "message": "Exactly one file is supported per job; multi-file uploads are not supported.",
          "request_id": "req-1234abcd"
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ]
          },
          "queue_position": {
            "type": "integer",
            "nullable": true
          },
          "message": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgress"
          },
          "input": {
            "$ref": "#/components/schemas/JobInput"
          },
          "output": {
            "$ref": "#/components/schemas/JobOutput"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "status",
          "queue_position",
          "message",
          "created_at",
          "progress",
          "input",
          "output",
          "error"
        ]
      },
      "JobProgress": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "processed": {
            "type": "integer"
          },
          "errors": {
            "type": "integer"
          },
          "step": {
            "type": "string"
          }
        },
        "required": [
          "total",
          "processed",
          "errors",
          "step"
        ]
      },
      "JobInput": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "4",
              "2",
              "both"
            ]
          },
          "model": {
            "type": "string"
          },
          "job_label": {
            "type": "string",
            "nullable": true,
            "description": "Display/log metadata only; not part of job identity."
          },
          "file": {
            "$ref": "#/components/schemas/JobInputFile"
          }
        },
        "required": [
          "mode",
          "model",
          "job_label",
          "file"
        ]
      },
      "JobInputFile": {
        "type": "object",
        "properties": {
          "original_filename": {
            "type": "string"
          },
          "stored_name": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer"
          },
          "sha256": {
            "type": "string"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobModeResult"
            }
          }
        },
        "required": [
          "original_filename",
          "stored_name",
          "size_bytes",
          "sha256"
        ]
      },
      "JobModeResult": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "4",
              "2"
            ]
          },
          "signature": {
            "type": "string"
          },
          "cache_hit": {
            "type": "boolean"
          },
          "rate_seconds_per_second": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "error": {
            "type": "string"
          }
        }
      },
      "JobOutput": {
        "type": "object",
        "properties": {
          "ready": {
            "type": "boolean"
          },
          "artifact_name": {
            "type": "string",
            "nullable": true
          },
          "content_type": {
            "type": "string",
            "nullable": true
          },
          "size_bytes": {
            "type": "integer",
            "nullable": true
          },
          "signature": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "ready"
        ],
        "description": "Metadata for downloadable ZIP output. Artifact layout uses deterministic top-level folders (`all/`, `vocals/`)."
      }
    }
  }
}